+++ front/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@headlessui/react": "^1.7.17",
     "@tanstack/react-query": "^4.29.25",
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^13.0.0",
