+++ package.json
@@ -1,5 +1,6 @@
 {
   "dependencies": {
+    "@headlessui/react": "^1.7.17",
     "react-slick": "^0.29.0",
     "recoil": "^0.7.7",
     "slick-carousel": "^1.8.1"
